Update README.md
A new three-dimensional implementation of slope stability regarding the column distribution that discretizes the slip surface is presented in this thesis. This was conducted by examining a computational model consisting of the columns that were generated by different widths. The main interest was to observe the effect of this nonuniformity of the columns in the factor of safety calculations.